Absolute Value Removal & More Dynamic Mesh Testing
Removed unnecessary absolute values that were applied to a magnitude. Added multiplier for mesh direction that will probably get removed later for being unnecessary. Issues with dynamic meshes seem to come from the fact the collider is convex. Distance from points ranges from ~0.1 to ~208 on tested model, but distance must be set to at least ~206.5 for any attempt at deforming to be made, so values under that must be caused from vertices moving.