Position Relevant Deforming
Made deforming affected by the position of the object. Deforming used to assume that the object was based at (0, 0, 0) and would not deform properly if the position was moved. Deforming will now happen on the object on the position the collision takes place. Rotation is not accounted for, so deformation still has issues related to the non-rotated calculations.